Split out project loading capabilities from rust-analyzer crate
